move menu components to new folder
