+++ client/src/pages/account/MyMenusTab.js
@@ -2,7 +2,7 @@ import React, {useEffect, useState} from "react";
 import axios from "axios";
 import {FaCheck, FaPlus} from "react-icons/fa";
 import {FaX} from "react-icons/fa6";
-import MenuCard from "../../components/Admin/MenuCard";
+import MenuCard from "../../components/Menus/MenuCard";
 import "../../format/MenuCard.css"
 import "../../format/Tabs.css"
